+++ Interfaces/DTO/Slot.cs
@@ -15,7 +15,7 @@ namespace Interfaces.DTO
         public int? breakdown_id { get; set; }
         public string? breakdown_name { get; set; } 
         public int? breakdown_warranty { get; set; }
-        public int? cost {  get; set; }
+        public double? cost {  get; set; }
 
         public int mechanic_id { get; set; }
